Move FuzzyDict tests to separate file, enable Codacy coverage again.
