+++ pywinauto/fuzzydict.py
@@ -163,64 +163,3 @@ class FuzzyDict(dict):
                 format(str(lookfor), str(key), ratio))
 
         return item
-
-
-if __name__ == '__main__':
-    import unittest
-
-    class FuzzyTestCase(unittest.TestCase):
-
-        """Perform some tests"""
-
-        test_dict = {
-            'Hiya'  : 1,
-            'hiy\xe4' : 2,
-            'test3' : 3,
-            1: 324}
-
-        def test_creation_empty(self):
-            """Verify that not specifying any values creates an empty dictionary"""
-            fd = FuzzyDict()
-
-            self.assertEquals(fd, {})
-
-        def test_creation_dict(self):
-            """Test creating a fuzzy dict"""
-            fd = FuzzyDict(self.test_dict)
-            self.assertEquals(fd, self.test_dict)
-            self.assertEquals(self.test_dict['Hiya'], fd['hiya'])
-
-            fd2 = FuzzyDict(self.test_dict, cutoff = .8)
-            self.assertEquals(fd, self.test_dict)
-            self.assertRaises(KeyError, fd2.__getitem__, 'hiya')
-
-        def test_contains(self):
-            """Test checking if an item is in a FuzzyDict"""
-            fd = FuzzyDict(self.test_dict)
-
-            self.assertEquals(True, fd.__contains__('hiya'))
-
-            self.assertEquals(True, fd.__contains__('test3'))
-
-            self.assertEquals(True, fd.__contains__('hiy\xe4'))
-
-            self.assertEquals(False, fd.__contains__('FuzzyWuzzy'))
-
-            self.assertEquals(True, fd.__contains__(1))
-
-            self.assertEquals(False, fd.__contains__(23))
-
-        def test_get_item(self):
-            """Test getting items from a FuzzyDict"""
-            fd = FuzzyDict(self.test_dict)
-
-            self.assertEquals(self.test_dict["Hiya"], fd['hiya'])
-            self.assertRaises(KeyError, fd.__getitem__, 'FuzzyWuzzy')
-
-            fd2 = FuzzyDict(self.test_dict, cutoff = .14)
-
-            self.assertEquals(1, fd2['FuzzyWuzzy'])
-            self.assertEquals(324, fd2[1])
-            self.assertRaises(KeyError, fd2.__getitem__, 23)
-
-    unittest.main()
